# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OraProjekt API",
    "description": "Multi-tenant timesheet management API",
    "version": "1.0.0",
    "contact": {
      "name": "OraProjekt",
      "url": "https://oraprojekt.com"
    }
  },
  "servers": [
    { "url": "https://oraprojekt.com/api", "description": "Production" },
    { "url": "https://staging.oraprojekt.com/api", "description": "Staging" },
    { "url": "http://localhost:3000/api", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "op_session"
      }
    },
    "schemas": {
...
      "TenantDomain": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tenantId": { "type": "string" },
          "domain": { "type": "string" },
          "status": { "type": "string", "enum": ["PENDING", "ACTIVE", "REJECTED"] },
          "verifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
  "paths": {
...
    "/tenant-domains": {
      "get": {
        "summary": "List tenant domains (manager only)",
        "responses": {
          "200": {
            "description": "List of domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "domains": { "type": "array", "items": { "$ref": "#/components/schemas/TenantDomain" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add tenant domain",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["domain"],
                "properties": {
                  "domain": { "type": "string", "example": "kompania-ks.com" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Domain added (PENDING status)" },
          "409": { "description": "Domain already claimed" }
        }
      }
    }
  }
}